requirements.txt
# Google Cloud Dependencies
google-cloud-firestore==2.14.0
google-cloud-storage==2.14.0
google-api-python-client==2.111.0
google-auth==2.25.2
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0

# Configuration and Environment
python-dotenv==1.0.0
PyYAML==6.0.1

# Utilities
requests==2.31.0

# Cloud Functions Framework (for deployment)
functions-framework==3.5.0
